+++ app/ui/main/components/timeline.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 """
-ClipFlow 时间线组件
+ClipFlowCut 时间线组件
 多轨时间线编辑器：视频轨 / 音频轨 / 字幕轨
 """